Adds SM_PHORON_VALUE define
Adds SM_PHORON_VALUE define, which directly affects the value of phoron
in regards to supermatter and stuff. By default it is 20, this means 20
units of phoron, has the potential to increase the size of the
supermatter by 1.

This gets rid of "infinite phoron" and "infinite supermatter", making
both more rare.

Greatly reduces the rate at which supermatter shards can be grown, but
simultaneously greatly increases their value in the neutron furnace. I
believe this will make the neutron furnace a more viable tactic for
obtaining resources, as before, it was too much time, too little reward.
Now, for a size 10 shard, you should be able to produce 200 metal, or 10
bars of phoron.  Technically it's possible to only produce 11 bars
depending on how well configured the furnace is.